Major update.
Added a ton of stability. 
if the controller should disconnect, it should be able to re connect again without any major issues.
if for some reason the network goes down though there is no reconnection to the wifi, but i moved the connection into a function so it should be easy enough to reconnect by adding something as simple as a ping to the main loop every so often.